Upgrade dependencies
Upgrade dependencies to avoid dependency to the older version
go-libp2p-core which breaks the build in upstream projects that depend
on go-cid versions < `0.2.0`.

The journey to upgrade this dependency led to the upgrade of other
dependencies such as replacing `go-ipfs-config` with `go-ipfs/config`.

In the absence of tests add a CI job to build everything to improve
confidence in code changes.